refactor: 내 정보 네이밍 통일
