+++ frontend/src/pages/index.ts
@@ -1,7 +1,7 @@
 export { default as Main } from './Main';
 export { default as Detail } from './Detail';
 export { default as Create } from './Create';
-export { default as MyInfo } from './MyInfo';
+export { default as MyInformation } from './MyInformation';
 export { default as Auth } from './Auth';
 export { default as MyGroup } from './MyGroup';
 export { default as NotFound } from './NotFound';
